+++ workspace/packages/muse-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ebay/muse-webpack-plugin
 
+## 1.0.29
+
+### Patch Changes
+
+- Enhance local plugins support.
+
 ## 1.0.28
 
 ### Patch Changes